requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
openai
firebase-admin
google-cloud-firestore